extract crypto utils
