+++ libs/importer/src/importers/lastpass/lib/parser.ts
@@ -4,6 +4,7 @@ import { Utils } from "@bitwarden/common/platform/misc/utils";
 import { Account } from "./account";
 import { BinaryReader } from "./binaryReader";
 import { Chunk } from "./chunk";
+import { CryptoUtils } from "./cryptoUtils";
 import { ParserOptions } from "./parserOptions";
 import { SharedFolder } from "./sharedFolder";
 
@@ -16,9 +17,11 @@ const AllowedSecureNoteTypes = new Set<string>([
 
 export class Parser {
   cryptoFunctionService: CryptoFunctionService;
+  cryptoUtils: CryptoUtils;
 
   constructor(cryptoFunctionService: CryptoFunctionService) {
     this.cryptoFunctionService = cryptoFunctionService;
+    this.cryptoUtils = new CryptoUtils(cryptoFunctionService);
   }
 
   /*
@@ -42,14 +45,14 @@ export class Parser {
     const id = Utils.fromBufferToUtf8(this.readItem(reader));
 
     // 1: name
-    const name = await this.decryptAes256PlainWithDefault(
+    const name = await this.cryptoUtils.decryptAes256PlainWithDefault(
       this.readItem(reader),
       encryptionKey,
       placeholder
     );
 
     // 2: group
-    const group = await this.decryptAes256PlainWithDefault(
+    const group = await this.cryptoUtils.decryptAes256PlainWithDefault(
       this.readItem(reader),
       encryptionKey,
       placeholder
@@ -66,7 +69,7 @@ export class Parser {
     }
 
     // 4: extra (notes)
-    const notes = await this.decryptAes256PlainWithDefault(
+    const notes = await this.cryptoUtils.decryptAes256PlainWithDefault(
       this.readItem(reader),
       encryptionKey,
       placeholder
@@ -79,14 +82,14 @@ export class Parser {
     this.skipItem(reader);
 
     // 7: username
-    let username = await this.decryptAes256PlainWithDefault(
+    let username = await this.cryptoUtils.decryptAes256PlainWithDefault(
       this.readItem(reader),
       encryptionKey,
       placeholder
     );
 
     // 8: password
-    let password = await this.decryptAes256PlainWithDefault(
+    let password = await this.cryptoUtils.decryptAes256PlainWithDefault(
       this.readItem(reader),
       encryptionKey,
       placeholder
@@ -214,7 +217,7 @@ export class Parser {
     this.skipItem(reader);
 
     // 39: totp (?)
-    const totp = await this.decryptAes256PlainWithDefault(
+    const totp = await this.cryptoUtils.decryptAes256PlainWithDefault(
       this.readItem(reader),
       encryptionKey,
       placeholder
@@ -265,7 +268,7 @@ export class Parser {
 
     // Name
     const encryptedName = this.readItem(reader);
-    const name = await this.decryptAes256Base64(encryptedName, key);
+    const name = await this.cryptoUtils.decryptAes256Base64(encryptedName, key);
 
     const folder = new SharedFolder();
     folder.id = id;
@@ -275,7 +278,7 @@ export class Parser {
   }
 
   async parseEncryptedPrivateKey(encryptedPrivateKey: string, encryptionKey: Uint8Array) {
-    const decrypted = await this.decryptAes256(
+    const decrypted = await this.cryptoUtils.decryptAes256(
       Utils.fromHexToArray(encryptedPrivateKey),
       encryptionKey,
       "cbc",
@@ -355,88 +358,5 @@ export class Parser {
 
   private readPayload(reader: BinaryReader, size: number): Uint8Array {
     return reader.readBytes(size);
-  }
-
-  // Crypto utils
-
-  private async decryptAes256PlainWithDefault(
-    data: Uint8Array,
-    encryptionKey: Uint8Array,
-    defaultValue: string
-  ) {
-    try {
-      return this.decryptAes256Plain(data, encryptionKey);
-    } catch {
-      return defaultValue;
-    }
-  }
-
-  private async decryptAes256Base64WithDefault(
-    data: Uint8Array,
-    encryptionKey: Uint8Array,
-    defaultValue: string
-  ) {
-    try {
-      return this.decryptAes256Base64(data, encryptionKey);
-    } catch {
-      return defaultValue;
-    }
-  }
-
-  private async decryptAes256Plain(data: Uint8Array, encryptionKey: Uint8Array) {
-    if (data.length === 0) {
-      return "";
-    }
-    // Byte 33 == character '!'
-    if (data[0] === 33 && data.length % 16 === 1 && data.length > 32) {
-      return this.decryptAes256CbcPlain(data, encryptionKey);
-    }
-    return this.decryptAes256EcbPlain(data, encryptionKey);
-  }
-
-  private async decryptAes256Base64(data: Uint8Array, encryptionKey: Uint8Array) {
-    if (data.length === 0) {
-      return "";
-    }
-    // Byte 33 == character '!'
-    if (data[0] === 33) {
-      return this.decryptAes256CbcBase64(data, encryptionKey);
-    }
-    return this.decryptAes256EcbBase64(data, encryptionKey);
-  }
-
-  private async decryptAes256EcbPlain(data: Uint8Array, encryptionKey: Uint8Array) {
-    return this.decryptAes256(data, encryptionKey, "ecb");
-  }
-
-  private async decryptAes256EcbBase64(data: Uint8Array, encryptionKey: Uint8Array) {
-    const d = Utils.fromB64ToArray(Utils.fromBufferToUtf8(data));
-    return this.decryptAes256(d, encryptionKey, "ecb");
-  }
-
-  private async decryptAes256CbcPlain(data: Uint8Array, encryptionKey: Uint8Array) {
-    const d = data.subarray(17);
-    const iv = data.subarray(1, 17);
-    return this.decryptAes256(d, encryptionKey, "cbc", iv);
-  }
-
-  private async decryptAes256CbcBase64(data: Uint8Array, encryptionKey: Uint8Array) {
-    const d = Utils.fromB64ToArray(Utils.fromBufferToUtf8(data.subarray(26)));
-    const iv = Utils.fromB64ToArray(Utils.fromBufferToUtf8(data.subarray(1, 25)));
-    return this.decryptAes256(d, encryptionKey, "cbc", iv);
-  }
-
-  private async decryptAes256(
-    data: Uint8Array,
-    encryptionKey: Uint8Array,
-    mode: "cbc" | "ecb",
-    iv: Uint8Array = new Uint8Array(16)
-  ): Promise<string> {
-    if (data.length === 0) {
-      return "";
-    }
-    // TODO: pass mode
-    const plain = await this.cryptoFunctionService.aesDecrypt(data, iv, encryptionKey);
-    return Utils.fromBufferToUtf8(plain);
   }
 }
